requirements.txt
langchain
langchain-openai
langchain-community
python-dotenv
numexpr
langchainhub
pep8